Added before adding identity

+++ Backend/TrelloClone/Repository/BoardRepository.cs
@@ -12,12 +12,10 @@ namespace TrelloClone.Repository
     {
 
         private readonly DataContext _context;
-        private readonly IMapper _mapper;
 
-        public BoardRepository(DataContext context, IMapper mapper)
+        public BoardRepository(DataContext context)
         {
             _context = context;
-            _mapper = mapper;
         }
 
         // SHOULD BE DONE IN BOARD SERVICE, VIA MEMBERSHIP SERVICE.
@@ -26,43 +24,49 @@ namespace TrelloClone.Repository
             throw new NotImplementedException();
         }*/
 
-        public async Task<KanbanBoardShortDTO> CreateBoard(string title, string username)
+        public async Task<KanbanBoard> CreateBoard(string title)
         {
 
-            KanbanBoard newBoard = new KanbanBoard();
-            newBoard.Title = title;
+            KanbanBoard newBoard = new() { Title = title };
+
             //newBoard.Memberships.Add(new Membership() { BoardId = newBoard.Id, UserId = username });
 
             await _context.Boards.AddAsync(newBoard);
             await _context.SaveChangesAsync();
 
-            return _mapper.Map<KanbanBoardShortDTO>(newBoard);
+            return newBoard;
         }
 
         public async Task DeleteBoard(int boardid)
         {
-            var board = await _context.Boards.FirstAsync(b => b.Id == boardid);
-            _context.Boards.Remove(board);
-            await _context.SaveChangesAsync();
+            var board = await _context.Boards.FindAsync(boardid);
+            if (board != null)
+            {
+                _context.Boards.Remove(board);
+                await _context.SaveChangesAsync();
+            }
         }
 
-        public async Task<ICollection<KanbanBoardShortDTO>> GetAllBoards()
+        public async Task<ICollection<KanbanBoard>> GetAllBoards()
         {
 
             var boards = await _context.Boards
-                .ProjectTo<KanbanBoardShortDTO>(_mapper.ConfigurationProvider)
+                .Include(b => b.Memberships).ThenInclude(m => m.User)
+                .Include(b => b.BoardLists).ThenInclude(bList => bList.Cards)
                 .ToListAsync();
 
 
             return boards;
         }
 
-        public async Task<KanbanBoardShortDTO> GetBoard(int boardid)
+        public async Task<KanbanBoard> GetBoard(int boardid)
         {
             var board = await _context.Boards
-                .FindAsync(boardid);
+                .Include(b => b.Memberships).ThenInclude(m => m.User)
+                .Include(b => b.BoardLists).ThenInclude(bList => bList.Cards)
+                .FirstAsync(b => b.Id == boardid);
 
-            return _mapper.Map<KanbanBoardShortDTO>(board);
+            return board;
 
 
         }
@@ -73,16 +77,13 @@ namespace TrelloClone.Repository
             return boardExists;
         }
 
-        public async Task<KanbanBoardShortDTO> UpdateBoard(int newBoardId, string newTitle)
+        public async Task<KanbanBoard> UpdateBoard(KanbanBoard updatedBoard)
         {
-            var oldBoard = await _context.Boards.Where(b => b.Id == newBoardId).SingleAsync();
 
-            oldBoard.Title = newTitle;
-
-            _context.Boards.Update(oldBoard);
+            _context.Boards.Update(updatedBoard);
             await _context.SaveChangesAsync();
 
-            return _mapper.Map<KanbanBoardShortDTO>(oldBoard);
+            return updatedBoard;
         }
     }
 }
